🐛 Fix:Post 생성 시 userId 포함되지 않는 문제 해결

+++ backend/src/post/post.service.ts
@@ -1,16 +1,24 @@
-import { Injectable } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  ForbiddenException,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { Model } from 'mongoose';
+import { Model, Types } from 'mongoose';
 import { Post } from './entities/post.schema';
 import { CreatePostDto } from './dto/create-post.dto';
 import { UpdatePostDto } from './dto/update-post.dto';
+import { UserDocument } from '../users/schemas/user.schema'; // UserDocument 임포트 추가
 
 @Injectable()
 export class PostService {
   constructor(@InjectModel(Post.name) private postModel: Model<Post>) {}
 
-  async create(createPostDto: CreatePostDto): Promise<Post> {
-    const createdPost = new this.postModel(createPostDto);
+  async create(createPostDto: CreatePostDto, userId: string): Promise<Post> {
+    const createdPost = new this.postModel({
+      ...createPostDto,
+      userId, // 게시글 작성자 ID 설정
+    });
     return await createdPost.save();
   }
 
@@ -19,16 +27,34 @@ export class PostService {
   }
 
   async findOne(id: string): Promise<Post> {
-    return await this.postModel.findById(id).exec();
+    const post = await this.postModel.findById(id).exec();
+    if (!post) {
+      throw new NotFoundException('게시글을 찾을 수 없습니다.');
+    }
+    return post;
   }
 
-  async update(id: string, updatePostDto: UpdatePostDto): Promise<Post> {
+  async update(
+    id: string,
+    updatePostDto: UpdatePostDto,
+    userId: string,
+  ): Promise<Post> {
+    const post = await this.findOne(id); // 게시글 찾기
+    if (post.userId.toString() !== userId) {
+      // 권한 체크
+      throw new ForbiddenException('본인의 게시글만 수정할 수 있습니다.');
+    }
     return await this.postModel
       .findByIdAndUpdate(id, updatePostDto, { new: true })
       .exec();
   }
 
-  async remove(id: string): Promise<any> {
+  async remove(id: string, userId: string): Promise<any> {
+    const post = await this.findOne(id); // 게시글 찾기
+    if (post.userId.toString() !== userId) {
+      // 권한 체크
+      throw new ForbiddenException('본인의 게시글만 삭제할 수 있습니다.');
+    }
     return await this.postModel.findByIdAndDelete(id).exec();
   }
 }
